requirements.txt
torch
torchvision
torchmetrics
wandb
imageio[ffmpeg]
einops
lpips
xformers
peft==0.9.0
diffusers==0.25.1
huggingface-hub==0.25.1
transformers==4.38.0
hydra-core>=1.3
hydra-submitit-launcher>=1.2